Feat: make new cube by Konva
