+++ package.json
@@ -8,10 +8,12 @@
     "fabric": "^4.6.0",
     "firebase": "^9.1.0",
     "gsap": "^3.8.0",
+    "konva": "^8.2.1",
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-gsap": "^3.2.1",
+    "react-konva": "^17.0.2-5",
     "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
     "styled-components": "^5.3.1",
